requirements.txt
streamlit
langchain
langchain-core
langchain-openai
python-dotenv